remove debug message
