+++ service/controller/app/v1/resource/chart/resource.go
@@ -2,15 +2,10 @@ package chart
 
 import (
 	"context"
-	"fmt"
-
-	v1alpha1 "github.com/giantswarm/apiextensions/pkg/apis/application/v1alpha1"
 	"github.com/giantswarm/apiextensions/pkg/clientset/versioned"
 	"github.com/giantswarm/microerror"
 	"github.com/giantswarm/micrologger"
 	"github.com/giantswarm/operatorkit/controller"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 
 	"github.com/giantswarm/app-operator/service/controller/app/v1/key"
@@ -41,60 +36,7 @@ type Resource struct {
 }
 
 func (r *Resource) GetCurrentState(ctx context.Context, obj interface{}) (interface{}, error) {
-	fmt.Println("it's created!!!")
 	return nil, nil
-}
-
-func (r *Resource) GetDesiredState(ctx context.Context, obj interface{}) (interface{}, error) {
-	fmt.Println("it's desired state!!!")
-	customResource, err := key.ToCustomResource(obj)
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	catalogName := key.CatalogName(customResource)
-
-	appCatalog, err := r.g8sClient.ApplicationV1alpha1().AppCatalogs("default").Get(catalogName, v1.GetOptions{})
-	if err != nil {
-		return nil, microerror.Mask(err)
-	}
-
-	chartURL := generateCatalogURL(appCatalog.Spec.CatalogStorage.URL, customResource.Spec.Name, customResource.Spec.Release)
-
-	chartCR := &v1alpha1.Chart{
-		TypeMeta: metav1.TypeMeta{
-			Kind:       chartKind,
-			APIVersion: chartAPIVersion,
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        customResource.Spec.Name,
-			Labels:      customResource.GetObjectMeta().GetLabels(),
-			Annotations: customResource.GetObjectMeta().GetAnnotations(),
-		},
-		Spec: v1alpha1.ChartSpec{
-			Name:       customResource.GetObjectMeta().GetName(),
-			Namespace:  customResource.Spec.Namespace,
-			TarballURL: chartURL,
-		},
-	}
-
-	if customResource.Spec.KubeConfig != (v1alpha1.AppSpecKubeConfig{}) {
-		chartCR.Spec.KubeConfig.Secret.Name = customResource.Spec.KubeConfig.Secret.Name
-		chartCR.Spec.KubeConfig.Secret.Namespace = customResource.Spec.KubeConfig.Secret.Namespace
-	}
-
-	if customResource.Spec.Config != (v1alpha1.AppSpecConfig{}) {
-		chartCR.Spec.Config.Secret.Name = customResource.Spec.Config.Secret.Name
-		chartCR.Spec.Config.Secret.Namespace = customResource.Spec.Config.Secret.Namespace
-
-		chartCR.Spec.Config.ConfigMap.Name = customResource.Spec.Config.ConfigMap.Name
-		chartCR.Spec.Config.ConfigMap.Namespace = customResource.Spec.Config.ConfigMap.Namespace
-	}
-	return chartCR, nil
-}
-
-func generateCatalogURL(baseURL string, appName string, release string) string {
-	return fmt.Sprintf("%s-%s-%s.tgz", baseURL, appName, release)
 }
 
 func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*controller.Patch, error) {
@@ -105,11 +47,24 @@ func (r *Resource) NewUpdatePatch(ctx context.Context, obj, currentState, desire
 }
 
 func (r *Resource) NewDeletePatch(ctx context.Context, obj, currentState, desiredState interface{}) (*controller.Patch, error) {
-	fmt.Println("it's patched!!!")
 	return nil, nil
 }
 
 func (r *Resource) ApplyCreateChange(ctx context.Context, obj, createChange interface{}) error {
+	app, err := key.ToCustomResource(obj)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+	chart, err := key.ToChart(createChange)
+	if err != nil {
+		return microerror.Mask(err)
+	}
+
+	_, err = r.g8sClient.ApplicationV1alpha1().Charts(app.Namespace).Create(&chart)
+
+	if err != nil {
+		return microerror.Mask(err)
+	}
 	return nil
 }
